allow multiple shared mem sources to recover after partial reads
The shared mem source froze in some instances because the connection to
shared memory was lost after at least one, but not all of the shm sources
had read a frame and incremented the position. The next read then always failed
because the shm sources had different positions and couldn't be synchronized
because they aren't seekable. This commit allows seek() to be called and just
updates the position, which is ok because it is a live source and position
is just an indicator of successful reads rather than a real position.